requirements (1).txt
face_recognition==1.3.0
facenet_pytorch==2.5.2
Flask==1.1.4
Flask_Cors==3.0.8
Flask_SQLAlchemy==2.5.1
imutils==0.5.4
numpy==1.17.0
opencv_python==4.2.0.34
PyJWT==2.3.0
scikit_learn==1.0.2
SQLAlchemy==1.3.23
torch==1.11.0+cu113
torchvision==0.12.0+cu113
Werkzeug==1.0.1
